refactore - Check and clean all the project from unusual code and components
