+++ src/components/mainNav/style.css
@@ -18,14 +18,9 @@
 }
 
 .nav-list__link:hover {
-  /* color: var(--light-blue-200); */
   opacity: 0.6;
   transition: color 0.5s ease-in;
 }
-
-/* .nav-list__link:focus {
-  color: var(--light-blue-300);
-} */
 
 .nav-list__link::before {
   content: '';
@@ -70,8 +65,6 @@
 .nav-list__active {
   transform: translateX(0);
 }
-
-/* Toggle burger menu line animation*/
 
 .toggle .line1 {
   transform: rotate(-45deg) translate(-4px, 5px);
