--- src/components/pricesList/style.css
@@ -1,63 +1,0 @@
-.prices__list {
-  display: flex;
-  justify-content: space-between;
-}
-
-.prices__list-item {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  width: 100%;
-  max-width: 35rem;
-  flex-grow: 0;
-  -ms-flex-negative: 0;
-  flex-shrink: 0;
-  min-height: 15.75rem;
-  padding: 4rem 2rem 1.9375rem;
-  border: 2px solid var(--light-blue-100);
-  border-radius: var(--border-rs);
-  box-sizing: border-box;
-}
-
-.prices__list-item-wrap {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-.prices__list-item-header {
-  min-height: 78px;
-  max-width: 354px;
-  margin: 0 0 0.5rem;
-  font-size: 2.25rem;
-  line-height: 2.6rem;
-  font-weight: 500;
-  letter-spacing: 0.02em;
-  color: var(--light-blue-100);
-}
-
-.prices__list-item-text {
-  font-size: 1.8rem;
-  line-height: 1.6;
-  color: var(--light-blue-100);
-}
-
-.prices__list-item-price {
-  display: inline-block;
-  margin-top: auto;
-  font-size: 2.9rem;
-  font-weight: 600;
-  line-height: 3rem;
-  color: var(--light-blue-300);
-}
-
-@media screen and (max-width: 1224px) {
-  .prices__list {
-    flex-direction: column;
-    gap: 4rem;
-  }
-
-  .prices__list-item {
-    max-width: 100%;
-  }
-}
