+++ src/components/scrollToTop/ScrollToTop.jsx
@@ -5,7 +5,6 @@ const ScrollToTop = () => {
   const { pathname } = useLocation();
 
   useLayoutEffect(() => {
-    // Мгновенный скролл до загрузки контента
     window.scrollTo(0, 0);
   }, [pathname]);
 
@@ -13,18 +12,3 @@ const ScrollToTop = () => {
 };
 
 export default ScrollToTop;
-
-// import { useEffect } from 'react';
-// import { useLocation } from 'react-router-dom';
-
-// const ScrollToTop = () => {
-//   const { pathname } = useLocation();
-
-//   useEffect(() => {
-//     window.scrollTo(0, 0);
-//   }, [pathname]);
-
-//   return null;
-// };
-
-// export default ScrollToTop;
